+++ backend/src/controllers/cdogs.js
@@ -7,6 +7,19 @@ const useCommonComponents = require("./useCommonComponents/index.js");
  * @returns {object}
  */
 
-const controller = useCommonComponents("/api/v2", "cdogs");
+const controller = useCommonComponents("cdogs");
+
+/**
+ * Get a dictionary of supported input template file types and output file types.
+ *
+ * @param   {FastifyRequest} request FastifyRequest is an instance of the standard http or http2 request objects.
+ * @param   {FastifyReply}   reply   FastifyReply is an instance of the standard http or http2 reply types.
+ * @returns {object}
+ */
+controller.getFileTypes = async (request, reply) => {
+  // Using Axios to call api endpoint with Bearer token
+  const response = await controller.api.get("/fileTypes");
+  return response;
+};
 
 module.exports = controller;